miniupnpd: update to 1.9.20141209

+++ trunk/user/miniupnpd/miniupnpd-1.9.x/testasyncsendto.c
@@ -1,4 +1,4 @@
-/* $Id: testasyncsendto.c,v 1.2 2014/02/25 11:00:14 nanard Exp $ */
+/* $Id: testasyncsendto.c,v 1.3 2014/11/28 17:08:56 nanard Exp $ */
 /* MiniUPnP project
  * http://miniupnp.free.fr/ or http://miniupnp.tuxfamily.org/
  * (c) 2006-2014 Thomas Bernard
@@ -72,7 +72,7 @@ int test(void)
 		struct timeval timeout;
 		struct timeval now;
 		syslog(LOG_DEBUG, "get_next_scheduled_send : %d next_send=%ld.%06ld",
-		       i, next_send.tv_sec, next_send.tv_usec);
+		       i, (long)next_send.tv_sec, (long)next_send.tv_usec);
 		FD_ZERO(&writefds);
 		max_fd = 0;
 		gettimeofday(&now, NULL);
@@ -98,7 +98,7 @@ int test(void)
 		}
 		syslog(LOG_DEBUG, "get_sendto_fds() returned %d", i);
 		syslog(LOG_DEBUG, "select(%d, NULL, xx, NULL, %ld.%06ld)",
-		       max_fd, timeout.tv_sec, timeout.tv_usec);
+		       max_fd, (long)timeout.tv_sec, (long)timeout.tv_usec);
 		i = select(max_fd, NULL, &writefds, NULL, &timeout);
 		if(i < 0) {
 			syslog(LOG_ERR, "select: %m");
